Fix chance of Mask of Endugu granting nerfed Iron Maiden skill
There are now 3 different types of Mask of Endugu that can spawn - all identical, but with different skills ranges:

 - 6-75 (69 skills - vanilla)
 - 77-155 (78 skills - vanilla)
 - 221-280 (59 skills - expansion)

The nerfed Iron Maiden (skill ID 76) can no longer be spawned.

Blackhorn's Face has been made 3x more likely to spawn relative to other unique Death Masks, to account for the 3 different flavours of Mask of Endugu.

Sadly, a Mask of Endugu cannot spawn with the actual Iron Maiden skill (skill ID 358). Adding a new item just for this one skill is not really worthwhile, as the rarities of all unique Death Masks would have to be skewed horribly to make it drop with the correct chance (negligible).